refactor(zim): extract _ContentMixin to zim/content.py
Move entry-content reading methods (get_zim_entry, _get_entry_content,
_get_entry_content_direct, _get_zim_entry_from_archive, get_entries,
get_binary_entry, _format_binary_response, _format_size,
get_entry_summary, _extract_entry_summary, _extract_html_summary, and
_get_entry_snippet) into _ContentMixin.

ZimOperations now inherits from _SearchMixin and _ContentMixin.
zim_archive and MAX_REDIRECT_DEPTH are accessed via the zim_operations
module so existing test patches keep working.
